cortex/python/requirements.txt
lightrag-hku
fastapi
uvicorn
networkx
numpy

flask-cors
google-generativeai
python-dotenv
pyautogui
websockets
websockets
requests
torch
transformers
accelerate
bitsandbytes
pillow
huggingface_hub
protobuf
sentencepiece
piper-tts
soundfile
